\name{pause}
\alias{pause}

\title{ Pause functions where necessary. }
\description{
  Used where the user may need to stop and prepare for future steps in the \code{\link{CPGmenu}} function
}
\usage{
pause()
}
\value{
  Simply \code{Press any key to continue...}
  }
\author{Andrew Parnell }

\examples{
	pause()
}
\keyword{methods}
